refactor code 2
